Teach --recursive to submodule summary
The submodule summary command was somehow left out when
--recursive was added to the other submodule commands.

Teach summary to handle the --recursive switch by recursing
into any dirty submodules and executing cmd_summary there,
too.

Change the reports of submodule $name to include the path
from the top of the original superproject.  Avoid reporting an
empty summary for submodules we recurse into just to execute
the summary there but which do not have any new commits to
summarize.

Signed-off-by: Phil Hord <hordp@cisco.com>