submodule: Add public submodule odb extender
Add a function to submodule.c to extend the alternates
objects database (in-memory) to include a specified submodule
and optionally validate that a commit sha1 is reachable.

Wrap the existing function 'add_submodule_odb' with helpers
to lookup the sha1, to assemble the path from a prefix + path
and to interpret the results in the return value.

Define an enum of possible values for this function to return.

Teach get_tree_entry to recurse submodule trees natively.
WIP: Probably need to turn that behavior on with some sort
of "--recurse-submodules" flag.

Submodule ref lookups work to some degree.   Still need to correct
some TODOs.  git rev-parse HEAD:sub1/foo works now.